+++ egui_commonmark/examples/mixing.rs
@@ -81,7 +81,7 @@ item a4 | item b4
 
                         "#,
                     r#"
- ![Rust logo](egui_commonmark/examples/rust-logo-128x128.png)
+ ![Ferris](egui_commonmark/examples/cuddlyferris.png)
                         "#,
                     r#"
 [Link to repo](https://github.com/lampsitter/egui_commonmark)
@@ -101,6 +101,13 @@ Term 3
 
 :   Definition 3
                         "#,
+                    r#"
+<a>Inline html</a>"#,
+                    r#"
+<p>
+Html
+block
+</p>"#,
                 );
 
                 ui.label("Label!");