+++ egui_commonmark/src/lib.rs
@@ -24,6 +24,9 @@
 //! ```toml
 //! image = { version = "0.25", default-features = false, features = ["png"] }
 //! ```
+//! # FAQ
+//!
+//! ## URL is not displayed when hovering over a link
 //!
 //! By default egui does not show urls when you hover hyperlinks. To enable it,
 //! you can do the following before calling any ui related functions:
@@ -75,6 +78,8 @@ mod parsers;
 
 pub use egui_commonmark_backend::alerts::{Alert, AlertBundle};
 pub use egui_commonmark_backend::misc::CommonMarkCache;
+pub use egui_commonmark_backend::RenderHtmlFn;
+pub use egui_commonmark_backend::RenderMathFn;
 
 #[cfg(feature = "macros")]
 pub use egui_commonmark_macros::*;
@@ -87,11 +92,11 @@ pub use egui_commonmark_backend;
 use egui_commonmark_backend::*;
 
 #[derive(Debug, Default)]
-pub struct CommonMarkViewer {
-    options: CommonMarkOptions,
+pub struct CommonMarkViewer<'f> {
+    options: CommonMarkOptions<'f>,
 }
 
-impl CommonMarkViewer {
+impl<'f> CommonMarkViewer<'f> {
     pub fn new() -> Self {
         Self::default()
     }
@@ -165,6 +170,55 @@ impl CommonMarkViewer {
     /// are used
     pub fn alerts(mut self, alerts: AlertBundle) -> Self {
         self.options.alerts = alerts;
+        self
+    }
+
+    /// Allows rendering math. This has to be done manually as you might want a different
+    /// implementation for the web and native.
+    ///
+    /// The example is template code for rendering a svg image. Make sure to enable the
+    /// `egui_extras/svg` feature for the result to show up.
+    ///
+    /// ## Example
+    ///
+    /// ```
+    /// # use std::{cell::RefCell, collections::HashMap, rc::Rc, sync::Arc};
+    /// # use egui_commonmark::CommonMarkViewer;
+    /// let mut math_images = Rc::new(RefCell::new(HashMap::new()));
+    /// CommonMarkViewer::new()
+    ///     .render_math_fn(Some(&move |ui, math, inline| {
+    ///         let mut map = math_images.borrow_mut();
+    ///         let svg = map
+    ///             .entry(math.to_string())
+    ///             .or_insert_with(|| {
+    ///                 if inline {
+    ///                     // render as inline
+    ///                     // dummy data for the example
+    ///                     Arc::new([0])
+    ///                 } else {
+    ///                     Arc::new([0])
+    ///                 }
+    ///             });
+    ///
+    ///     let uri = format!("{}.svg", egui::Id::from(math.to_string()).value());
+    ///     ui.add(
+    ///          egui::Image::new(egui::ImageSource::Bytes {
+    ///             uri: uri.into(),
+    ///             bytes: egui::load::Bytes::Shared(svg.clone()),
+    ///          })
+    ///          .fit_to_original_size(1.0),
+    ///     );
+    ///     }));
+    /// ```
+    pub fn render_math_fn(mut self, func: Option<&'f RenderMathFn>) -> Self {
+        self.options.math_fn = func;
+        self
+    }
+
+    /// Allows custom handling of html. Enabling this will disable plain text rendering
+    /// of html blocks. Nodes are included in the provided text
+    pub fn render_html_fn(mut self, func: Option<&'f RenderHtmlFn>) -> Self {
+        self.options.html_fn = func;
         self
     }