+++ egui_commonmark_backend/src/misc.rs
@@ -18,8 +18,7 @@ const DEFAULT_THEME_LIGHT: &str = "base16-ocean.light";
 #[cfg(feature = "better_syntax_highlighting")]
 const DEFAULT_THEME_DARK: &str = "base16-ocean.dark";
 
-#[derive(Debug)]
-pub struct CommonMarkOptions {
+pub struct CommonMarkOptions<'f> {
     pub indentation_spaces: usize,
     pub max_image_width: Option<usize>,
     pub show_alt_text_on_hover: bool,
@@ -33,9 +32,35 @@ pub struct CommonMarkOptions {
     pub alerts: AlertBundle,
     /// Whether to present a mutable ui for things like checkboxes
     pub mutable: bool,
+    pub math_fn: Option<&'f crate::RenderMathFn>,
+    pub html_fn: Option<&'f crate::RenderHtmlFn>,
 }
 
-impl Default for CommonMarkOptions {
+impl<'f> std::fmt::Debug for CommonMarkOptions<'f> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let mut s = f.debug_struct("CommonMarkOptions");
+
+        s.field("indentation_spaces", &self.indentation_spaces)
+            .field("max_image_width", &self.max_image_width)
+            .field("show_alt_text_on_hover", &self.show_alt_text_on_hover)
+            .field("default_width", &self.default_width);
+
+        #[cfg(feature = "better_syntax_highlighting")]
+        s.field("theme_light", &self.theme_light)
+            .field("theme_dark", &self.theme_dark);
+
+        s.field("use_explicit_uri_scheme", &self.use_explicit_uri_scheme)
+            .field(
+                "default_implicit_uri_scheme",
+                &self.default_implicit_uri_scheme,
+            )
+            .field("alerts", &self.alerts)
+            .field("mutable", &self.mutable)
+            .finish()
+    }
+}
+
+impl Default for CommonMarkOptions<'_> {
     fn default() -> Self {
         Self {
             indentation_spaces: 4,
@@ -50,11 +75,13 @@ impl Default for CommonMarkOptions {
             default_implicit_uri_scheme: "file://".to_owned(),
             alerts: AlertBundle::gfm(),
             mutable: false,
+            math_fn: None,
+            html_fn: None,
         }
     }
 }
 
-impl CommonMarkOptions {
+impl CommonMarkOptions<'_> {
     #[cfg(feature = "better_syntax_highlighting")]
     pub fn curr_theme(&self, ui: &Ui) -> &str {
         if ui.style().visuals.dark_mode {
@@ -199,7 +226,7 @@ pub struct Image {
 impl Image {
     // FIXME: string conversion
     pub fn new(uri: &str, options: &CommonMarkOptions) -> Self {
-        let has_scheme = uri.contains("://");
+        let has_scheme = uri.contains("://") || uri.starts_with("data:");
         let uri = if options.use_explicit_uri_scheme || has_scheme {
             uri.to_string()
         } else {
@@ -342,6 +369,7 @@ impl CodeBlock {
 fn simple_highlighting(ui: &Ui, text: &str, extension: &str) -> egui::text::LayoutJob {
     egui_extras::syntax_highlighting::highlight(
         ui.ctx(),
+        ui.style(),
         &egui_extras::syntax_highlighting::CodeTheme::from_style(ui.style()),
         text,
         extension,
@@ -543,6 +571,9 @@ pub fn prepare_show(cache: &mut CommonMarkCache, ctx: &egui::Context) {
         // cache free from egui's Ui and Context types as this allows it to be created before
         // any egui instances. It also keeps the API similar to before the introduction of the
         // image loaders.
+        #[cfg(feature = "embedded_image")]
+        crate::data_url_loader::install_loader(ctx);
+
         egui_extras::install_image_loaders(ctx);
         cache.has_installed_loaders = true;
     }